# docs/swagger.yaml
basePath: /
definitions:
...
  domain.AuthResponse:
    properties:
      access_token:
        type: string
      patient:
        $ref: '#/definitions/domain.Patient'
      refresh_token:
        type: string
    type: object
...
  domain.ChangePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        minLength: 8
        type: string
    required:
    - current_password
    - new_password
    type: object
...
  domain.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  domain.InitializePaymentRequest:
    properties:
      appointment_id:
        type: string
    required:
    - appointment_id
    type: object
  domain.InitializePaymentResponse:
    properties:
      authorization_url:
        type: string
      payment:
        $ref: '#/definitions/domain.Payment'
      reference:
        type: string
    type: object
  domain.LoginDto:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  domain.Patient:
    properties:
      address:
        type: string
      created_at:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      gender:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_email_verified:
        type: boolean
      name:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
  domain.Payment:
    properties:
      amount:
        type: number
      appointment_id:
        type: string
      created_at:
        type: string
      currency:
        type: string
      id:
        type: string
      paid_at:
        type: string
      patient_id:
        type: string
      paystack_access_code:
        type: string
      paystack_reference:
        type: string
      status:
        description: pending, successful, failed, refunded
        type: string
      updated_at:
        type: string
    type: object
  domain.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
      user_id:
        type: string
    required:
    - refresh_token
    - user_id
    type: object
...
  domain.RegisterPatientDTO:
    properties:
      address:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      gender:
        enum:
        - male
        - female
        - other
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        minLength: 8
        type: string
      phone:
        maxLength: 20
        minLength: 10
        type: string
    required:
    - date_of_birth
    - email
    - gender
    - name
    - password
    - phone
    type: object
  domain.ResendOTPRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  domain.ResetPasswordRequest:
    properties:
      email:
        type: string
      new_password:
        minLength: 8
        type: string
      otp:
        type: string
    required:
    - email
    - new_password
    - otp
    type: object
...
  domain.VerifyEmailRequest:
    properties:
      email:
        type: string
      otp:
        type: string
    required:
    - email
    - otp
    type: object
host: localhost:8080
info:
  contact:
    email: your-email@gmail.com
    name: Emerald
  description: A production-grade healthcare API for Nigeria — appointment booking,
    medical records, and telemedicine
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: HealthConnect API
  version: "1.0"
paths:
...
  /api/v1/auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Sends a password reset OTP to the patient's registered email address.
      parameters:
      - description: Patient email for password reset.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Request patient password reset
      tags:
      - Authentication (Patient)
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a patient with their credentials and returns access
        and refresh tokens.
      parameters:
      - description: Patient login credentials.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.LoginDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "429":
          description: Too Many Requests
          schema:
            additionalProperties: true
            type: object
      summary: Login as a patient
      tags:
      - Authentication (Patient)
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Issues a new access token using a valid refresh token for patient.
      parameters:
      - description: Refresh token data.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Refresh patient access token
      tags:
      - Authentication (Patient)
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new patient account and sends an OTP to their email for
        verification.
      parameters:
      - description: Patient registration data.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.RegisterPatientDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created - Patient registered successfully.
          schema:
            $ref: '#/definitions/domain.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
      summary: Register a new patient
      tags:
      - Authentication (Patient)
  /api/v1/auth/resend-otp:
    post:
      consumes:
      - application/json
      description: Resends the email verification OTP to the patient's email.
      parameters:
      - description: Patient email to resend OTP to.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.ResendOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Resend patient verification OTP
      tags:
      - Authentication (Patient)
  /api/v1/auth/reset-password:
    post:
      consumes:
      - application/json
      description: Resets the patient's password using the OTP sent to their email.
      parameters:
      - description: Reset password data including email, OTP, and new password.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Reset patient password with OTP
      tags:
      - Authentication (Patient)
  /api/v1/auth/verify-email:
    post:
      consumes:
      - application/json
      description: Verifies a patient's email address using the OTP sent during registration
        or resend.
      parameters:
      - description: Email and OTP for verification.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Verify patient email with OTP
      tags:
      - Authentication (Patient)
...
  /api/v1/patients/change-password:
    post:
      consumes:
      - application/json
      description: Changes the authenticated patient's password. Requires current
        and new passwords.
      parameters:
      - description: Change password data.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - Patients
  /api/v1/patients/logout:
    post:
      description: Invalidates the patient's refresh token, effectively logging them
        out.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Logout patient
      tags:
      - Patients
  /api/v1/patients/me:
    get:
      description: Returns the authenticated patient's profile information.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Patient'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get current patient profile
      tags:
      - Patients
  /api/v1/payments/{appointment_id}:
    get:
      description: Returns the payment status for an appointment
      parameters:
      - description: Appointment ID
        in: path
        name: appointment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Payment'
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get payment status
      tags:
      - payments
  /api/v1/payments/initialize:
    post:
      consumes:
      - application/json
      description: Creates a Paystack payment link for the patient to pay
      parameters:
      - description: Payment data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.InitializePaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.InitializePaymentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Initialize payment for an appointment
      tags:
      - payments
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and your JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"